Support extended offsets for PCI config access

+++ include/kernel/pci.h
@@ -40,12 +40,12 @@ uint32_t pci_config_read_segment(uint16_t segment,
                                  uint8_t bus,
                                  uint8_t device,
                                  uint8_t function,
-                                 uint8_t offset);
+                                 uint16_t offset);
 void pci_config_write_segment(uint16_t segment,
                               uint8_t bus,
                               uint8_t device,
                               uint8_t function,
-                              uint8_t offset,
+                              uint16_t offset,
                               uint32_t value);
 
 static inline uint32_t pci_config_read(uint8_t bus,
