+++ src/kernel/drivers/bus/pciroot/pci.c
@@ -76,7 +76,7 @@ uint32_t pci_config_read_segment(uint16_t segment,
                                  uint8_t bus,
                                  uint8_t device,
                                  uint8_t function,
-                                 uint8_t offset) {
+                                 uint16_t offset) {
   pci_mmconfig_window_t* window = pci_mmconfig_find(segment, bus);
   if(window != NULL) {
     uintptr_t base = (uintptr_t)window->virt_base;
@@ -100,7 +100,7 @@ void pci_config_write_segment(uint16_t segment,
                               uint8_t bus,
                               uint8_t device,
                               uint8_t function,
-                              uint8_t offset,
+                              uint16_t offset,
                               uint32_t value) {
   pci_mmconfig_window_t* window = pci_mmconfig_find(segment, bus);
   if(window != NULL) {
